Essentially created whole thing but forgot to commit
Created a .section that contains a flex box with title and body in one column and an image in the other, the order is easily changable by adding -reversed onto class name in the html. Images in the section currently utilise object-fit:cover; which isn't supported in IE11 <, will probably come up with a fallback but is fine for now and works in chrome/edge/firefox. Added in all of the desired sections as per the designs on paper. Also added in a footer which scales and changes flex-direction at 850px width.