Fixed img size and 4k font size
Changed all of the image's sizes to be 500x300, this works exactly the same with object-fit (even faster loading times) but also means browsers that don't support object-fit don't break the page, just results in aspect ratio distortion at small screen sizes (which is fine for stuff like IE which isn't used on mobile). Also set font size to 24-32px at 4k or around 4k resolutions so that it is still readable.